requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-extra-fields==3.0.2
djangorestframework==3.11.1
gunicorn==20.0.4
import-logger==1.0.0
Pillow==7.2.0
psycopg2-binary==2.8.6
PyPDF2==1.26.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
